+++ CHANGES.md
@@ -4,6 +4,10 @@ CHANGES
 unreleased
 ----------
 * Fixed the use of == to compare Strings in FormElement. See #49
+* Fixed 'Get Alert Message'. See #50
+  * 'Get Alert Message' does not confirm the Alert any more. 'Get Alert Message' + Confirm = 'Confirm Action'.
+  * Added keyword 'Choose Cancel On Confirmation'
+  * Added keyword 'Choose Ok On Confirmation'
 
 1.4.0.4
 -------